proxy-api/requirements.txt
# Foundational Stack
fastapi~=0.115.0
uvicorn[standard]~=0.34.0
pydantic~=2.10.0
pydantic-settings~=2.7.0

# LLM Routing
litellm~=1.56.0

# HTTP Client (Presidio calls)
httpx~=0.28.0

# Structured Logging
structlog~=24.4.0

# Config
pyyaml~=6.0.0

# Development/Testing
pytest~=8.3.0
pytest-asyncio~=0.24.0
